pyproject.toml
[project]
name = "benign-attention"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"accelerate>=1.6.0",
	"datasets>=3.5.1",
	"einops>=0.8.1",
	"hydra-core>=1.3.2",
	"jupyter>=1.1.1",
	"markupsafe==2.1.5",
	"matplotlib>=3.10.1",
	"pandas>=2.2.3",
	"seaborn>=0.13.2",
	"torch>=2.6.0",
	"torchaudio>=2.6.0",
	"torchvision>=0.21.0",
	"transformers>=4.51.3",
	"wandb>=0.19.9",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D1",  # undocumented
]
unfixable = [
    "F401",  # unused document
    "F841",  # unused variable
]
